pyproject.toml
[project]
name = "changes-metadata-manager"
version = "1.0.0"
description = "Map RDF knowledge graph triples to Aldrovandi digitization folder stages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "rdflib>=7.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "master"
commit_message = "chore(release): {version}"

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.remote]
type = "github"
